Use RegExs to generate credit card information
We use the regexs to define more complex patterns and remove the need
for helper functions to pad strings or generate random values. We can
now also refactor out the regexs and reuse them in other libraries.
Adding a new card type is now quite simple. By using metaprogramming for
both the unit tests as well as the lib code we reduce the burden on the
developer to add new types.

This can be used as a template for other types of information. Many RFCs
come with regexs for the given standard types. This should allow for
quick additions of new types.

Added libraries:
  * RandExp : ^0.4.2